Temporarily set ssh-keygen to use old PKCS#1 PEM format
SDC and its users have a lot of scripts and tools which expect
ssh-keygen to produce the old private key format.

We want to balance this against the fact that the new format
vastly improves security for passphrase protected keys, so if
we are generating a key with a passphrase, we will default
to the new format, otherwise old.